+++ Core/WhatsNew.lua
@@ -74,6 +74,8 @@ CHANGELOG = {
             L["wn_404_layout_share_safety"],
             L["wn_404_interface_reliability"],
             L["wn_404_studio_return"],
+            L["wn_404_actionbar_glow_color"],
+            L["wn_404_floating_combat_text"],
             L["wn_404_release_validation"],
             L["wn_404_secret_reference"],
             L["wn_404_secret_audit"],